Scala Native support for http4sServer
