+++ server/http4s-server/src/main/scala/sttp/tapir/server/http4s/Http4sServerInterpreter.scala
@@ -9,7 +9,6 @@ import org.http4s._
 import org.http4s.headers.`Content-Length`
 import org.http4s.server.websocket.WebSocketBuilder2
 import org.http4s.websocket.WebSocketFrame
-import org.log4s.{Logger, getLogger}
 import org.typelevel.ci.CIString
 import sttp.capabilities.WebSockets
 import sttp.capabilities.fs2.Fs2Streams
@@ -112,8 +111,6 @@ trait Http4sServerInterpreter[F[_]] {
 }
 
 object Http4sServerInterpreter {
-
-  private[http4s] val log: Logger = getLogger
 
   def apply[F[_]]()(implicit _fa: Async[F]): Http4sServerInterpreter[F] = {
     new Http4sServerInterpreter[F] {
