+++ server/http4s-server/src/main/scala/sttp/tapir/server/http4s/Http4sServerOptions.scala
@@ -32,21 +32,7 @@ object Http4sServerOptions {
 
   def defaultDeleteFile[F[_]](implicit sync: Sync[F]): TapirFile => F[Unit] = file => sync.blocking(Defaults.deleteFile()(file))
 
-  def defaultServerLog[F[_]: Sync]: DefaultServerLog[F] = {
-    DefaultServerLog(
-      doLogWhenReceived = debugLog(_, None),
-      doLogWhenHandled = debugLog[F],
-      doLogAllDecodeFailures = debugLog[F],
-      doLogExceptions = (msg: String, ex: Throwable) => Sync[F].delay(Http4sServerInterpreter.log.error(ex)(msg)),
-      noLog = Sync[F].pure(())
-    )
-  }
-
-  private def debugLog[F[_]](msg: String, exOpt: Option[Throwable])(implicit sync: Sync[F]): F[Unit] =
-    exOpt match {
-      case None     => Sync[F].delay(Http4sServerInterpreter.log.debug(msg))
-      case Some(ex) => Sync[F].delay(Http4sServerInterpreter.log.debug(ex)(msg))
-    }
+  def defaultServerLog[F[_]: Sync]: DefaultServerLog[F] = Http4sDefaultServerLogOptions.defaultServerLog[F]
 
   def default[F[_]: Sync]: Http4sServerOptions[F] = customiseInterceptors[F].options
 }
